requirements.txt
asgiref==3.4.1
Deprecated==1.2.13
Django==3.2.9
django-cacheops==6.0
djangorestframework==3.12.4
funcy==1.17
packaging==21.3
Pillow==9.0.0
pyparsing==3.0.6
python-dotenv==0.19.2
pytz==2021.3
redis==4.1.0
six==1.16.0
sqlparse==0.4.2
wrapt==1.13.3
